backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
websockets>=12.0
aiohttp>=3.9.0
python-multipart>=0.0.6
pyyaml>=6.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.10.0

# Code execution
RestrictedPython==8.1

# Database
aiosqlite>=0.19.0
